Data cleanup rewrite

+++ app/Helpers/Delete/Models/DeleteRevenuesKwh.php
@@ -9,8 +9,11 @@
 namespace App\Helpers\Delete\Models;
 
 
+use App\Eco\Cooperation\Cooperation;
 use App\Helpers\Delete\DeleteInterface;
+use Carbon\Carbon;
 use Illuminate\Database\Eloquent\Model;
+use Illuminate\Support\Facades\Log;
 
 /**
  * Class DeleteRevenueDistribution
@@ -19,9 +22,12 @@ use Illuminate\Database\Eloquent\Model;
  */
 class DeleteRevenuesKwh implements DeleteInterface
 {
-
+    private bool $force = true; // default: oude gedrag (hard)
     private $errorMessage = [];
     private $revenuesKwh;
+    private $yearsForDelete;
+    private $dateAllowedToDelete;
+    private $cooperation;
 
     /** Sets the model to delete
      *
@@ -30,8 +36,33 @@ class DeleteRevenuesKwh implements DeleteInterface
     public function __construct(Model $revenuesKwh)
     {
         $this->revenuesKwh = $revenuesKwh;
+        $this->cooperation = Cooperation::first();
+        $cleanupItemRevenueKwh = $this->cooperation->cleanupItems()->where('code_ref', 'revenuesKwh')->first();
+        $this->yearsForDelete = $cleanupItemRevenueKwh?->years_for_delete ?? 99;
+        $this->dateAllowedToDelete = Carbon::now()->subYears($this->yearsForDelete)->format('Y-m-d');
+
     }
 
+    /** If it's called by the cleanup functionality, we land on this function, else on the delete function
+     *
+     * @return array
+     * @throws
+     */
+    public function cleanup()
+    {
+        try{
+            $this->force = false;   // cleanup = soft
+            return $this->delete();
+        }catch (\Exception $exception){
+            Log::error('Fout bij opschonen Opbrengsten Kwh', [
+                'exception' => $exception->getMessage(),
+                'errormessages' => implode(' | ', $this->errorMessage),
+            ]);
+            array_push($this->errorMessage, "Fout bij opschonen Opbrengsten Kwh. (meld dit bij Econobis support)");
+            return $this->errorMessage;
+        }
+
+    }
     /** Main method for deleting this model and all it's relations
      *
      * @return array errorMessage array
@@ -39,13 +70,16 @@ class DeleteRevenuesKwh implements DeleteInterface
      */
     public function delete()
     {
-        $this->canDelete();
+        if (! $this->canDelete()) {
+            return $this->errorMessage;
+        }
         $this->deleteModels();
         $this->dissociateRelations();
         $this->deleteRelations();
         $this->customDeleteActions();
-        if( !sizeof($this->errorMessage)>0 ) {
-            $this->revenuesKwh->forceDelete();
+        if (count($this->errorMessage) === 0) {
+            $this->force ? $this->revenuesKwh->forceDelete()
+                : $this->revenuesKwh->delete();
         }
 
         return $this->errorMessage;
@@ -53,26 +87,44 @@ class DeleteRevenuesKwh implements DeleteInterface
 
     /** Checks if the model can be deleted and sets error messages
      */
-    public function canDelete()
+    public function canDelete(): bool
     {
-        if($this->revenuesKwh->confirmed){
-            array_push($this->errorMessage, "Opbrengstverdeling Kwh is al definitief.");
-            return;
+        // indien status processed en einddatum revenue ligt voor bewaarplicht termijn, dan ok
+        if($this?->revenuesKwh?->confirmed && $this?->revenuesKwh?->status === 'processed' && $this?->revenuesKwh?->date_end < $this->dateAllowedToDelete) {
+            return true;
         }
+        // indien status processed en einddatum revenue ligt op of na bewaarplicht termijn, dan niet ok
+        if($this?->revenuesKwh?->confirmed && $this?->revenuesKwh?->status === 'processed' && $this?->revenuesKwh?->date_end >= $this->dateAllowedToDelete) {
+            array_push($this->errorMessage, "Er is al een verwerkte opbrengstverdeling Kwh aangemaakt. Opbrengstverdeling Kwh kan niet worden verwijderd vanwege de bewaarplicht: " . $this->yearsForDelete . " jaar.");
+            return false;
+        }
+        // overige situaties:
+        // indien revenue bevestigd (confirmed), dan niet ok
+        if($this->revenuesKwh->confirmed){
+            array_push($this->errorMessage, "Opbrengstverdeling is al definitief.");
+            return false;
+        }
+        return true;
     }
 
     /** Deletes models recursive
      */
     public function deleteModels()
     {
-        foreach($this->revenuesKwh->distributionKwh as $distributionKwh) {
-                $deleteRevenueDistributionKwh = new DeleteRevenueDistributionKwh($distributionKwh);
-                $this->errorMessage = array_merge($this->errorMessage, ( $deleteRevenueDistributionKwh->delete() ?? [] ) );
+        $queryDistributionKwh = $this->revenuesKwh->distributionKwh();
+        $itemsDistributionKwh = $this->force ? $queryDistributionKwh->withTrashed()->get() : $queryDistributionKwh->get();
+        foreach ($itemsDistributionKwh as $distributionKwh) {
+            $deleteRevenueDistributionKwh = new DeleteRevenueDistributionKwh($distributionKwh);
+            $this->errorMessage = array_merge( $this->errorMessage, ($this->force ? $deleteRevenueDistributionKwh->delete() : ($deleteRevenueDistributionKwh->cleanup() ?? [])));
         }
-        foreach($this->revenuesKwh->partsKwh as $partsKwh) {
-                $deleteRevenuePartsKwh = new DeleteRevenuePartsKwh($partsKwh);
-                $this->errorMessage = array_merge($this->errorMessage, ( $deleteRevenuePartsKwh->delete() ?? [] ) );
+
+        $queryPartsKwh = $this->revenuesKwh->partsKwh();
+        $itemsPartsKwh = $this->force ? $queryPartsKwh->withTrashed()->get() : $queryPartsKwh->get();
+        foreach ($itemsPartsKwh as $partsKwh) {
+            $deleteRevenuePartsKwh = new DeleteRevenuePartsKwh($partsKwh);
+            $this->errorMessage = array_merge( $this->errorMessage, ($this->force ? $deleteRevenuePartsKwh->delete() : ($deleteRevenuePartsKwh->cleanup() ?? [])));
         }
+
     }
 
 
@@ -93,7 +145,22 @@ class DeleteRevenuesKwh implements DeleteInterface
      */
     public function customDeleteActions()
     {
-        $this->revenuesKwh->conceptValuesKwh()->delete();
+        // indien status processed en einddatum revenue ligt voor bewaarplicht termijn
+        if($this?->revenuesKwh?->confirmed && $this?->revenuesKwh?->status === 'processed' && $this?->revenuesKwh?->date_end < $this->dateAllowedToDelete) {
+            if ($this->force) {
+                $this->revenuesKwh->valuesKwh()->withTrashed()->forceDelete();
+            } else {
+                $this->revenuesKwh->valuesKwh()->delete();
+            }
+        } else {
+            if ($this->force) {
+                $this->revenuesKwh->conceptValuesKwh()->withTrashed()->forceDelete();
+            } else {
+                $this->revenuesKwh->conceptValuesKwh()->delete();
+            }
+
+        }
+
     }
